improved database initialization
