+++ islands/ViewNote.tsx
@@ -14,6 +14,7 @@ import { ViewNoteSchema, viewNoteSchema } from '../lib/validation/note.ts';
 import * as v from '@valibot/valibot';
 import ErrorPage from '../components/ErrorPage.tsx';
 import PasswordToggle from '../components/PasswordToggle.tsx';
+import CopyButton from '../components/CopyButton.tsx';
 
 // Constants for messages
 const MESSAGES = {
@@ -269,13 +270,7 @@ function DisplayDecryptedNote(
                     {/* Content section */}
                     <CardContent>
                         <div class='flex items-center gap-2 mb-4'>
-                            <svg class='w-5 h-5 text-gray-300' fill='currentColor' viewBox='0 0 20 20'>
-                                <path
-                                    fill-rule='evenodd'
-                                    d='M4 4a2 2 0 012-2h4.586A2 2 0 0112 2.586L15.414 6A2 2 0 0116 7.414V16a2 2 0 01-2 2H6a2 2 0 01-2-2V4zm2 6a1 1 0 011-1h6a1 1 0 110 2H7a1 1 0 01-1-1zm1 3a1 1 0 100 2h6a1 1 0 100-2H7z'
-                                    clip-rule='evenodd'
-                                />
-                            </svg>
+                            <CopyButton value={content} />
                             <h3 class='text-xl font-semibold text-white'>Content</h3>
                         </div>
