+++ routes/api/notes.ts
@@ -2,28 +2,26 @@ import { createNoteSchema } from '../../lib/validation/note.ts';
 import { formatExpiration, Note } from '../../lib/types.ts';
 import * as v from '@valibot/valibot';
 import { Context } from 'fresh';
-import { getNoteDatabase } from '../../lib/services/database-service.ts';
 import * as bcrypt from 'bcrypt';
 import { State } from '../../lib/types/common.ts';
+import { noteDatabase } from '../../main.ts';
 
 /* used for client side note creation and encryption
-	* This endpoint handles only POST requests.
-	* - POST: Creates a new note with the provided content, IV, password, and expiration time.
-	*
-	* Note: The content should be encrypted before sending to this endpoint and the password should be hashed with PBKDF2
-	* on the client side for security, then securely hashed with bcrypt on the server for storage.
-	*
-	* rate-limiting (ARC - Anonymous Rate-Limited Credentials):
-	* - Limit: 10 requests per minute per client
-	* - Block duration: 5 minutes for rate limit violations
-	* - Privacy-preserving: Uses anonymous tokens with daily rotation
-	* - No IP address storage: Only hashed, rotated tokens are kept
-	*/
+    * This endpoint handles only POST requests.
+    * - POST: Creates a new note with the provided content, IV, password, and expiration time.
+    *
+    * Note: The content should be encrypted before sending to this endpoint and the password should be hashed with PBKDF2
+    * on the client side for security, then securely hashed with bcrypt on the server for storage.
+    *
+    * rate-limiting (ARC - Anonymous Rate-Limited Credentials):
+    * - Limit: 10 requests per minute per client
+    * - Block duration: 5 minutes for rate limit violations
+    * - Privacy-preserving: Uses anonymous tokens with daily rotation
+    * - No IP address storage: Only hashed, rotated tokens are kept
+    */
 
 export const handler = {
     async POST(ctx: Context<State>) {
-        const db = await getNoteDatabase();
-
         try {
             const { content, iv, password, expiresIn, manualDeletion } = await ctx.req.json();
 
@@ -42,7 +40,7 @@ export const handler = {
                 );
             }
 
-            const noteId = await db.generateNoteId();
+            const noteId = await noteDatabase.generateNoteId();
             const hasPassword = password && password.trim() !== '';
 
             // if password is provided, hash it with bcrypt (password should be PBKDF2 hashed on client before sending)
@@ -58,7 +56,7 @@ export const handler = {
                 manualDeletion: manualDeletion,
             };
 
-            const insertResult = await db.insertNote(result);
+            const insertResult = await noteDatabase.insertNote(result);
 
             if (!insertResult.success) {
                 return new Response(
