+++ routes/api/notes/[id].ts
@@ -1,14 +1,13 @@
 import { Note } from '../../../lib/types.ts';
 import { Context } from 'fresh';
-import { getNoteDatabase } from '../../../lib/services/database-service.ts';
 import * as bcrypt from 'bcrypt';
 import { State } from '../../../lib/types/common.ts';
+import { noteDatabase } from '../../../main.ts';
 
 export const handler = async (ctx: Context<State>): Promise<Response> => {
     if (ctx.req.method !== 'POST' && ctx.req.method !== 'DELETE') {
         return new Response('Method not allowed', { status: 405 });
     }
-    const db = await getNoteDatabase();
 
     const id = ctx.params.id;
     if (!id) {
@@ -16,7 +15,7 @@ export const handler = async (ctx: Context<State>): Promise<Response> => {
     }
 
     if (ctx.req.method === 'POST') {
-        const note = await db.getNoteById(id);
+        const note = await noteDatabase.getNoteById(id);
         const { passwordHash } = await ctx.req.json();
 
         if (!note || !passwordHash) {
@@ -29,7 +28,7 @@ export const handler = async (ctx: Context<State>): Promise<Response> => {
 
         // If the note doesn't require manual deletion, delete it to ensure it has been destroyed
         if (!note.manualDeletion) {
-            await db.deleteNote(id);
+            await noteDatabase.deleteNote(id);
         }
 
         return new Response(
@@ -45,7 +44,7 @@ export const handler = async (ctx: Context<State>): Promise<Response> => {
             },
         );
     } else if (ctx.req.method === 'DELETE') {
-        const note = await db.getNoteById(id);
+        const note = await noteDatabase.getNoteById(id);
         const { passwordHash } = await ctx.req.json();
         if (!note) {
             return new Response('Note not found', { status: 404 });
@@ -54,7 +53,7 @@ export const handler = async (ctx: Context<State>): Promise<Response> => {
         if (note.password && !compareHash(passwordHash, note.password)) {
             return new Response('Invalid password or auth key', { status: 403 });
         }
-        await db.deleteNote(id);
+        await noteDatabase.deleteNote(id);
         return new Response(
             JSON.stringify({
                 message: 'Note deleted successfully',
